Day 6
Write a program to find the Quadrants in which coordinates lie.

Get the value of x and y coordinates as input from the user and check in which quadrant the point lies and print it.

Input : 10  20
Output : This point lies in the first quadrant.

Input : 10 20
Output : This point lies in the second quadrant.