Day 8
Write a program to find roots of a quadratic equation.

Get the values of a, b and c (coefficients of quadratic equation) as input from the user and calculate the roots and print as the output.

Input:Enter the value of a, b and c : 1 -6 9
Output:
Roots are equal
Root 1= root 2 = 3.00